test: restore paired nested SSH fixture after HUB restart
